python/document-parser/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pypdf==5.9.0
pytesseract==0.3.13
pillow==11.3.0
pytest==8.4.1
httpx==0.28.1
